chore: Move prisma scripts to prisma/scripts/
