+++ server/prisma/scripts/nuke.ts
